Recustomize how power consumption is computed
Here is my setup on the electric box:
  [master] --> [stuff-1]
           --> [sub-master-1] --> [stuff-2]
           --> [sub-master-2] --> [stuff-3]

Since there are no single cable available to mesure stuff-1, I can't get
a direct measured value of stuff-1.

I can monitor sub-master-1, sub-master-2 and the master.

With all of theses information, 'stuff-1' is actually computable:
  stuff-1 = master - sub-master-1 - sub-master-2

This commit change how the frontend compute and show this. It's possible
that nothing is used on stuff-1 (actually this is the heating system).
It occures lot of time that substraction gives negative values. With
sensors precision, I usualy get around -10w and -50w. Let assume this is
a precision issue, if value is negative, just assume nothing was
consumed.

This commit, in the same time, change the month summary using a stacked
chart of the 3 channels (aka 'stuff-X') to display consumption of each
part.