Add new master/slave workflow in place of thraded
Previous version was using one single python process threaded with async
stuff to handle different parts of the monitoring stuff.

This is maybe good for resources saving but it comes with downside like
concurrency debugging and database locking.

This new approach is less sexy but provide more flexibility. A pub/sub
approach is used, with a master websocket server and a redis in between
which receive data from slaves and forward them to websocket clients.

The advantages of this approach is the dedication of slaves code not
needed to be attached to the master process and could be written in
other languages than python